requirements.txt
pypdf
langchain
chromadb # Vector storage
pytest
openai
langchain-community
langchain-openai
markdown
streamlit
openai-agents
python-dotenv
pydantic
serpapi
langchain.tools
langchain-chroma
pandas
json